move to bridge aggregators, avoid silent ignoring of errors
